rename servicetalk-opentelemetry -> servicetalk-opentelemetry-http
