+++ servicetalk-opentelemetry-http/src/test/java/io/servicetalk/opentelemetry/http/HeadersPropagatorGetterTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.servicetalk.opentelemetry;
+package io.servicetalk.opentelemetry.http;
 
 import io.servicetalk.http.api.DefaultHttpHeadersFactory;
 import io.servicetalk.http.api.HttpHeaders;
