+++ servicetalk-opentelemetry-http/src/test/java/io/servicetalk/opentelemetry/http/OpenTelemetryHttpServerFilterTest.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package io.servicetalk.opentelemetry;
+package io.servicetalk.opentelemetry.http;
 
 import io.servicetalk.http.api.HttpClient;
 import io.servicetalk.http.api.HttpResponse;
 import io.servicetalk.http.netty.HttpServers;
 import io.servicetalk.log4j2.mdc.utils.LoggerStringWriter;
-import io.servicetalk.opentelemetry.TestUtils.TestTracingServerLoggerFilter;
+import io.servicetalk.opentelemetry.http.TestUtils.TestTracingServerLoggerFilter;
 import io.servicetalk.transport.api.ServerContext;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -46,9 +46,9 @@ import java.net.URL;
 import static io.servicetalk.concurrent.api.Single.succeeded;
 import static io.servicetalk.http.netty.HttpClients.forSingleAddress;
 import static io.servicetalk.log4j2.mdc.utils.LoggerStringWriter.stableAccumulated;
-import static io.servicetalk.opentelemetry.OpenTelemetryHttpRequestFilterTest.verifyTraceIdPresentInLogs;
-import static io.servicetalk.opentelemetry.TestUtils.SPAN_STATE_SERIALIZER;
-import static io.servicetalk.opentelemetry.TestUtils.TRACING_TEST_LOG_LINE_PREFIX;
+import static io.servicetalk.opentelemetry.http.OpenTelemetryHttpRequestFilterTest.verifyTraceIdPresentInLogs;
+import static io.servicetalk.opentelemetry.http.TestUtils.SPAN_STATE_SERIALIZER;
+import static io.servicetalk.opentelemetry.http.TestUtils.TRACING_TEST_LOG_LINE_PREFIX;
 import static io.servicetalk.transport.netty.internal.AddressUtils.localAddress;
 import static io.servicetalk.transport.netty.internal.AddressUtils.serverHostAndPort;
 import static org.assertj.core.api.Assertions.assertThat;
