+++ servicetalk-opentelemetry-http/src/test/java/io/servicetalk/opentelemetry/http/TestSpanState.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.servicetalk.opentelemetry;
+package io.servicetalk.opentelemetry.http;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
