+++ settings.gradle
@@ -91,8 +91,8 @@ include "servicetalk-annotations",
         "servicetalk-logging-slf4j-internal",
         "servicetalk-oio-api",
         "servicetalk-oio-api-internal",
-        "servicetalk-opentelemetry",
         "servicetalk-opentelemetry-asynccontext",
+        "servicetalk-opentelemetry-http",
         "servicetalk-opentracing-inmemory",
         "servicetalk-opentracing-inmemory-api",
         "servicetalk-opentracing-internal",
